Split mouse and keyboard tests into individual files like in the code base
