+++ multisensor_pipeline/tests/test_keyboard.py
@@ -1,29 +1,6 @@
 from time import sleep
 
 
-def test_simple_mouse(xvfb):
-    from multisensor_pipeline.modules import QueueSink
-    from multisensor_pipeline.modules.mouse import Mouse
-    from multisensor_pipeline.pipeline.graph import GraphPipeline
-
-    # (1) define the modules
-    source = Mouse(move=True, scroll=True, click=True)
-    sink: QueueSink = QueueSink()
-
-    # (2) add module to a pipeline...
-    pipeline = GraphPipeline()
-    pipeline.add_source(source)
-    pipeline.add_sink(sink)
-    # (3) ...and connect the modules
-    pipeline.connect(source, sink)
-
-    pipeline.start()
-
-    sleep(.3)
-    pipeline.stop()
-    assert True
-
-
 def test_simple_keyboard(xvfb):
     from multisensor_pipeline.modules import QueueSink
     from multisensor_pipeline.modules.keyboard import Keyboard
@@ -43,6 +20,8 @@ def test_simple_keyboard(xvfb):
     pipeline.start()
     sleep(.3)
     pipeline.stop()
+
+    # If we ever get here, we consider this test successful.
     assert True
